chore: Update Haskell code to use command line arguments
This commit updates the Haskell code in `vocab1.hs` to include support for command line arguments. Previously, the code was expecting a single argument, but it didn't include the program name. With this change, the code now correctly handles command line arguments by using the `getArgs` function from the `System.Environment` module.

Additionally, this commit includes some comments to explain the purpose of the `dropAround` and `toCaseFold` functions used in the code.

The changes in this commit improve the usability and flexibility of the program by allowing users to pass command line arguments when running the program.